refactor(sdk)!: extract transport-free query core into dash-platform-queries
Split rs-sdk per the maintainer guidance to refactor rather than duplicate: the query-building, wire-encoding, and proof-decoding core that a transport-free embedder needs now lives in a new packages/dash-platform-queries crate, and rs-sdk depends on it and re-exports every moved item at its old path, so consumers keep their imports.

Moved out of rs-sdk: DocumentQuery and its wire encoders (document_query.rs), the count/sum/average/ranked proof helpers and their FromProof aggregate views (DocumentCount, DocumentSum, DocumentAverage, DocumentSplitCounts, DocumentSplitSums, DocumentSplitAverages, DocumentRankedEntries), DocumentHistoryQuery, block_info_from_metadata, FinalizedEpochQuery, ensure_valid_state_transition_structure, and the DPNS username helpers (convert_to_homograph_safe_chars, is_valid_username, is_contested_username). Sdk-bound pieces stay behind: the contract-fetching DocumentQuery constructor (now the DocumentQuerySdk extension trait), the Query<GetDocumentsRequest> encoder impl, the Fetch bindings for the aggregate views, and the Query impls for FinalizedEpochQuery.

QuerySettings stays in rs-sdk with its request_settings field intact: it is only ever consumed by rs-sdk's own Query::query implementations, so moving it bought nothing and would have removed a public field from a re-exported type.

block_info_from_metadata keeps an rs-sdk-owned forwarding function with its historical Result<BlockInfo, dash_sdk::Error> signature; the transport-free implementation is exposed separately for embedders.

The new crate has its own small thiserror enum (Config/Drive/Protocol); rs-sdk converts it via From, so existing ? call sites keep compiling. wasm-sdk gains the matching From impl for WasmSdkError, routed through SdkError so the mapping is unchanged.

Coherence fallout: with DocumentQuery now foreign to rs-sdk, the blanket 'impl Query<T> for T where T: TransportRequest' would conflict with the explicit identity impl for DocumentQuery. The blanket is now additionally bounded by a local, explicitly-implemented WireQuery marker covering every wire request proto (list mirrors rs-dapi-client's TransportRequest impls); rustc can then prove the impl sets disjoint.

Dependency graph: the new crate has no networking transport — no rs-dapi-client, hyper, rustls, tower, or tonic transport, asserted in CI alongside drive-proof-verifier. tokio and tonic's generated-code support remain reachable transitively (dash-context-provider -> dash-async, and dapi-grpc's client types), exactly as they already were for drive-proof-verifier; "transport-free" means no networking stack, not an async-runtime-free graph.

BREAKING CHANGE: three rs-sdk surfaces are source-incompatible for downstream crates.

1. DocumentQuery's own methods (new, try_into_request_for_version, the TryFromPlatformVersioned associated error, and TryFrom into DriveDocumentQuery) now yield dash_platform_queries::Error rather than dash_sdk::Error. dash_sdk::Error: From<dash_platform_queries::Error> keeps `?` call sites working; explicit return types, direct variant matching, and function-pointer/associated-error bounds need an SdkError::from conversion.

2. DocumentQuery::new_with_data_contract_id moved from an inherent method to the DocumentQuerySdk extension trait (an inherent impl is impossible from rs-sdk now that the type is foreign). Callers add `use dash_sdk::platform::DocumentQuerySdk;`.

3. The blanket `impl Query<T> for T` is additionally bounded by the new dash_sdk::platform::WireQuery marker. Every in-workspace request proto implements it; a downstream crate with its own TransportRequest type adds `impl dash_sdk::platform::WireQuery for MyRequest {}`.
