+++ packages/dash-platform-queries/src/documents/ranked_proof_helpers.rs
@@ -19,7 +19,7 @@
 //!
 //! [`DocumentRankedEntries`]: drive_proof_verifier::DocumentRankedEntries
 
-use crate::platform::documents::document_query::DocumentQuery;
+use crate::documents::document_query::DocumentQuery;
 use dapi_grpc::platform::v0::{GetDocumentsResponse, Proof, ResponseMetadata};
 use dapi_grpc::platform::VersionedGrpcResponse;
 use dash_context_provider::ContextProvider;
