+++ packages/dash-platform-queries/src/transition/mod.rs
@@ -1,0 +1,2 @@
+//! Transport-free state transition helpers.
+pub mod validation;
