+++ packages/dash-platform-queries/src/types/mod.rs
@@ -1,0 +1,2 @@
+//! Transport-free query types for various dpp objects.
+pub mod finalized_epoch;
