+++ packages/rs-sdk/src/platform.rs
@@ -32,6 +32,9 @@ pub use dash_context_provider::ContextProvider;
 pub use dash_context_provider::MockContextProvider;
 pub use documents::document_history_query::DocumentHistoryQuery;
 pub use documents::document_query::DocumentQuery;
+/// Sdk-bound constructors for [`DocumentQuery`]. Must be in scope to call
+/// [`DocumentQuery::new_with_data_contract_id`], which fetches the contract.
+pub use documents::DocumentQuerySdk;
 pub use dpp::{
     self as dpp,
     document::Document,
@@ -45,7 +48,7 @@ pub use {
     fetch_unproved::FetchUnproved,
     query::{
         IdentityKeysQuery, LimitQuery, ProposerBlockCountByIdsQuery, Query, QueryStartInfo,
-        RecentAddressBalanceChangesQuery, RecentCompactedAddressBalanceChangesQuery,
+        RecentAddressBalanceChangesQuery, RecentCompactedAddressBalanceChangesQuery, WireQuery,
         DEFAULT_EPOCH_QUERY_LIMIT,
     },
     query_settings::QuerySettings,
