+++ packages/rs-sdk/src/platform/documents/mod.rs
@@ -1,34 +1,18 @@
-pub(super) mod average_proof_helpers;
-pub(super) mod count_proof_helpers;
-/// `Fetch` impl for the average-side aggregate result. Returns
-/// `(count, sum)`; client divides.
-pub mod document_average;
-pub mod document_count;
-/// `Fetch` impl for the having-range (`GROUP BY … HAVING <aggregate>
-/// <op> <value> LIMIT n`) result — one entry per matching group, in
-/// axis order, with proof-attested completeness. Requires an index
-/// declaring `rankedCountable` / `rankedSummable` / `rankedAverageable`
-/// (protocol version 14+).
-pub mod document_having_entries;
-pub mod document_history_query;
-pub mod document_query;
-/// `Fetch` impl for the ranked (`GROUP BY … ORDER BY <aggregate> LIMIT n
-/// [OFFSET m]`) result — one entry per returned group, in ranking order,
-/// plus the rank the page starts at. Requires an index declaring
-/// `rankedCountable` / `rankedSummable` / `rankedAverageable`
-/// (protocol version 14+).
-pub mod document_ranked_entries;
-/// `Fetch` impl for the average-side per-entry result. Mirrors
-/// `document_split_sums`.
-pub mod document_split_averages;
-pub mod document_split_counts;
-/// `Fetch` impl for the sum-side per-entry result. Mirrors
-/// `document_split_counts`.
-pub mod document_split_sums;
-/// `Fetch` impl for the sum-side aggregate result. Mirrors
-/// `document_count`. Lights up alongside grovedb PR 670.
-pub mod document_sum;
-pub(super) mod having_proof_helpers;
-pub(super) mod ranked_proof_helpers;
-pub(super) mod sum_proof_helpers;
+//! Document query surface.
+//!
+//! The transport-free core (query types, wire encoding, proof decoding)
+//! lives in the `dash-platform-queries` crate and is re-exported here at
+//! its historical paths; this module keeps the Sdk-bound pieces — `Fetch`
+//! bindings, the contract-fetching constructor, and transition builders.
+
+pub use dash_platform_queries::documents::{
+    document_average, document_count, document_having_entries, document_history_query,
+    document_query, document_ranked_entries, document_split_averages, document_split_counts,
+    document_split_sums, document_sum,
+};
+
+pub mod document_query_sdk;
+mod fetch_bindings;
 pub mod transitions;
+
+pub use document_query_sdk::DocumentQuerySdk;
